[SYSML-282] Use "select" vector to remove empty rows/cols
This change was based on observation that time consumed by eviction was
high in experimental data. In few algorithms such as -- DecisionTree,
RandomForest -- in which Matrix was operated with RemoveEmpty operator.
Before doing RemoveEmpty operation, select matrix was operated on target
matrix with append and elementwise multiplication. This is repeated in
loop operation. Instead of operating select vector on target matrix,
select vector can be passed to RemoveEmpty to avoid expensive operations
causing delay in eviction.

This fix will provide an option for caller to send select vector in
RemoveEmpty operator.

Enhancement to RemoveEmpty include following:

- It will use "Select vector" if passed in to figure out rows/columns
  to be removed.
- Memory requirement estimate update.
- Test cases to evaluate this change updated.