[SYSML-311] Rework cp function call instructions (full operand info)
So far function call instructions contained the number and names of
inputs and outputs. In case of constant propagation of literals equal to
existing variable names, this led to a wrong mapping of input
parameters. We now compile the full operand information (name, data
type, value type, isliteral) of all inputs into function calls in order
to distinguish literals from existing variables. 

This change also includes (1) a new testcase for this function call
issue, (2) a fix for instruction compilation of multi-return builtin
functions with multiple inputs, and (3) a fix for integer overflows of
long default values.