[SYSML-317] Source function default current directory in PyDML
The default working directory was not being set to the
current directory in PyDML. Therefore, PydmlSyntacticValidator
is updated to match DmlSyntacticValidator in regards to
setting this directory.

Also updated SetWorkingDirTest to include PyDML tests
with accompanying *.pydml scripts. I placed these scripts
in the same directory as the *.dml scripts since currently
no other PyDML tests of this type exist.