[SYSML-320] Compiler decision on serialized caching (sparse matrices)
The in-memory representation of sparse or ultra-sparse matrix blocks is
significantly larger than their serialized form. This led to unnecessary
read per iteration if the in-memory representation does not fit in
aggregated memory. In order to overcome this bottleneck, we now
automatically decide on deserialized vs serialized storage levels. For
dense, we always use deserialized because the serialized form offers no
compression but has additional overhead. For sparse, we also use
deserialized if its estimated size fits into the total data memory;
otherwise we fall back to the serialized representation.