Improved cp left/right indexing (dense/sparse output), for kmeans
So far, we estimated the output sparsity and related representation for
left/right indexing operations with a uniformity assumption. The block
operation does not probe exam sparsity in order to prevent additional
memory requirements. However, in the presence of sparsity skew this led
to a wrong output representation which affected subsequent operations.
We now explicitly check the output dense/sparse representation after
unpinning the input matrix which guards the memory requirements of the
potential change of representations.