New multi-threaded vector-matrix block multiplications, incl tests
The default block matrix multiplications used a parallelization over
rows in the result because this allowed a simple parallelization w/
disjoint results. However, this approach did not cover a common case of
vector-matrix multiplications which often originates from a rewrite of
t(X)%*%y. This change introduces special cases for dense-dense,
dense-sparse, sparse-dense, sparse-sparse vector-matrix multiplication
with parallelization over the number of rows of the input matrix. With
this change, we now achieve also for vector-matrix mult, peak QPI memory
bandwidth of 25GB/s.