[SYSML-351] Enable unit tests in package functions.append to run in parallel
This applies the same pattern of changes made in SYSML-313 to package
com.ibm.bi.dml.test.integration.functions.append as described in document
attached to SYSML-286.

The package functions.append contains two relatively long tests
(AppendMatrix and AppendChain). When run as a suite, the set takes ~25 minutes
but when run in parallel ~15 minutes as mentioned in the 'Addtional Packages'
section of the document. Note the pom.xml change to disable the corresponding
Suite class will be handled in a separate PR to help minimize merge conflicts
as a few other PRs to be submitted to migrate other long-running packages.